+++ index.js
@@ -1,38 +1,14 @@
-const Crawler = require('./crawler');
-const crawler = new Crawler;
-
 const { IncomingWebhook } = require('@slack/client');
 const webhook = new IncomingWebhook(process.env.SLACK_WEBHOOK_URL);
 
-const createSlackMessage = () => {
-  let message = {
-    text: `最新の空き情報を取得しました！`,
-    mrkdwn: true,
-    attachments: [
-      {
-        title: `公園別空き情報一覧`,
-        title_link: process.env.ENDPOINT,
-        fields: [
-          {
-            title: `ほげほげ公園`,
-            value: `:ok:`
-          },
-          {
-            title: `ほげほげ海岸公園`,
-            value: `:ok:`
-          }
-        ]
-      }
-    ]
-  };
-  return message;
-}
+const crawler = require('./crawler');
+const createSlackMessage = require('./createSlackMessage');
 
 exports.courtCrawler = async (eventData, eventContext, callback) => {
   console.log(eventData);
   console.log(eventContext);
 
-  await crawler.start();
+  await crawler();
 
   webhook.send(createSlackMessage(), (err, res) => {
     if (err) {
