+++ test.js
@@ -1,7 +1,6 @@
 require('dotenv').config();
-const Crawler = require('./crawler');
+const crawler = require('./crawler');
 
 (async () => {
-  const crawler = new Crawler;
-  await crawler.start();
+  await crawler();
 })();
